indexing commit for last 24 hours of code

+++ src/dubstep/Main.java
@@ -36,7 +36,7 @@ public class Main {
             if (args[i].equals("--on-disk"))
                 mySchema.setInMem(false);
         }
-       mySchema.setInMem(true);
+        mySchema.setInMem(true);
 
         Scanner scanner = new Scanner(System.in);
         QueryTimer timer = new QueryTimer();
@@ -71,14 +71,40 @@ public class Main {
 //        executeQuery("create table R(id int,id1 int);");
 //        executeQuery("create table S(id int,id1 int);");
 
-//         executeQuery("CREATE TABLE LINEITEM(ORDERKEY INT,PARTKEY INT,SUPPKEY INT,LINENUMBER INT,QUANTITY DECIMAL,EXTENDEDPRICE DECIMAL,DISCOUNT DECIMAL,TAX DECIMAL,RETURNFLAG CHAR(1),LINESTATUS CHAR(1),SHIPDATE DATE,COMMITDATE DATE,RECEIPTDATE DATE,SHIPINSTRUCT CHAR(25),SHIPMODE CHAR(10),COMMENT VARCHAR(44),PRIMARY KEY (ORDERKEY,LINENUMBER));");
-//         executeQuery("CREATE TABLE ORDERS(ORDERKEY INT,CUSTKEY INT,ORDERSTATUS CHAR(1),TOTALPRICE DECIMAL,ORDERDATE DATE,ORDERPRIORITY CHAR(15),CLERK CHAR(15),SHIPPRIORITY INT,COMMENT VARCHAR(79),PRIMARY KEY (ORDERKEY));");
+         executeQuery("CREATE TABLE LINEITEM(ORDERKEY INT,PARTKEY INT,SUPPKEY INT,LINENUMBER INT,QUANTITY DECIMAL,EXTENDEDPRICE DECIMAL,DISCOUNT DECIMAL,TAX DECIMAL,RETURNFLAG CHAR(1),LINESTATUS CHAR(1),SHIPDATE DATE,COMMITDATE DATE,RECEIPTDATE DATE,SHIPINSTRUCT CHAR(25),SHIPMODE CHAR(10),COMMENT VARCHAR(44)," +
+                 "PRIMARY KEY (ORDERKEY,LINENUMBER), INDEX SHIPDATE (SHIPDATE), INDEX RECEIPTDATE (RECEIPTDATE), INDEX RETURNFLAG (RETURNFLAG));");
+         executeQuery("CREATE TABLE ORDERS(\n" +
+                 "  ORDERKEY INT, \n" +
+                 "  CUSTKEY INT, \n" +
+                 "  ORDERSTATUS CHAR(1), \n" +
+                 "  TOTALPRICE DECIMAL, \n" +
+                 "  ORDERDATE DATE, \n" +
+                 "  ORDERPRIORITY CHAR(15), \n" +
+                 "  CLERK CHAR(15), \n" +
+                 "  SHIPPRIORITY INT, \n" +
+                 "  COMMENT VARCHAR(79), \n" +
+                 "  PRIMARY KEY (ORDERKEY), \n" +
+                 "  INDEX ORDERDATE (ORDERDATE));");
+         //         executeQuery("CREATE TABLE SUPPLIER(\n" +
+//                 "  SUPPKEY INT, \n" +
+//                 "  NAME CHAR(25), \n" +
+//                 "  ADDRESS VARCHAR(40), \n" +
+//                 "  NATIONKEY INT, \n" +
+//                 "  PHONE CHAR(15), \n" +
+//                 "  ACCTBAL DECIMAL, \n" +
+//                 "  COMMENT VARCHAR(101), \n" +
+//                 "  PRIMARY KEY (SUPPKEY), \n" +
+//                 "  INDEX NATIONKEY (NATIONKEY));");
 //         executeQuery("CREATE TABLE PART(PARTKEY INT,NAME VARCHAR(55),MFGR CHAR(25),BRAND CHAR(10),TYPE VARCHAR(25),SIZE INT,CONTAINER CHAR(10),RETAILPRICE DECIMAL,COMMENT VARCHAR(23),PRIMARY KEY (PARTKEY));");
 //         executeQuery("CREATE TABLE CUSTOMER(CUSTKEY INT,NAME VARCHAR(25),ADDRESS VARCHAR(40),NATIONKEY INT,PHONE CHAR(15),ACCTBAL DECIMAL,MKTSEGMENT CHAR(10),COMMENT VARCHAR(117),PRIMARY KEY (CUSTKEY));");
 //         executeQuery("CREATE TABLE SUPPLIER(SUPPKEY INT,NAME CHAR(25),ADDRESS VARCHAR(40),NATIONKEY INT,PHONE CHAR(15),ACCTBAL DECIMAL,COMMENT VARCHAR(101),PRIMARY KEY (SUPPKEY));");
 //         executeQuery("CREATE TABLE PARTSUPP(PARTKEY INT,SUPPKEY INT,AVAILQTY INT,SUPPLYCOST DECIMAL,COMMENT VARCHAR(199),PRIMARY KEY (PARTKEY,SUPPKEY));");
 //         executeQuery("CREATE TABLE NATION(NATIONKEY INT,NAME CHAR(25),REGIONKEY INT,COMMENT VARCHAR(152),PRIMARY KEY (NATIONKEY));");
-//         executeQuery("CREATE TABLE REGION(REGIONKEY INT,NAME CHAR(25),COMMENT VARCHAR(152),PRIMARY KEY (REGIONKEY));");
+//         executeQuery("CREATE TABLE REGION(\n" +
+//                 "  REGIONKEY INT, \n" +
+//                 "  NAME CHAR(25), \n" +
+//                 "  COMMENT VARCHAR(152), \n" +
+//                 "  PRIMARY KEY (REGIONKEY));");
 
         while (scanner.hasNext()) {
 
@@ -109,14 +135,17 @@ public class Main {
 
         if (query instanceof CreateTable) {
             CreateTable createQuery = (CreateTable) query;
-            BufferedWriter table_file = null;
-            try {
-                table_file = new BufferedWriter(new FileWriter("tables",true));
+            File tableFile = new File("tables");
+            if (!tableFile.exists()) {
+                BufferedWriter table_file;
+                try {
+                    table_file = new BufferedWriter(new FileWriter("tables",true));
 
-                table_file.write(sqlString+"\n");
-                table_file.close();
-            } catch (IOException e) {
-                e.printStackTrace();
+                    table_file.write(sqlString+"\n");
+                    table_file.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
 
             if (!mySchema.createTable(createQuery)) {
